requirements.txt
fastapi
pydantic
google-generativeai>=0.7,<1.0
uvicorn
httpx
numpy
sqlalchemy
python-dotenv
alembic
asyncpg
faiss-cpu
pytest
pytest-asyncio>=0.23
httpx
aiosqlite
python-multipart
asgi-lifespan
